requirements.txt
decorator
pyyaml
plac
python-baseconv